Rename ComponentController to Controller and moved to controller module (is now used for non-component, e.g. SimulatorController)
