+++ vivarium/controllers/controller.py
@@ -1,31 +1,28 @@
 from operator import attrgetter
 
-from vivarium.controllers.dataclass_wrapper import Remote
-
 
 class AttributeMapping:
     def __init__(self, jax_attr, ctrl_attr=None, jax_to_ctrl_fn=None, ctrl_to_jax_fn=None):
         self.jax_attr = jax_attr
         self.ctrl_attr = ctrl_attr if ctrl_attr is not None else jax_attr
         self.jax_to_ctrl_fn = jax_to_ctrl_fn if jax_to_ctrl_fn is not None else lambda x: x
-        self.ctrl_to_jax_fn = ctrl_to_jax_fn if ctrl_to_jax_fn is not None else lambda x: x        
+        self.ctrl_to_jax_fn = ctrl_to_jax_fn if ctrl_to_jax_fn is not None else lambda x: x
+        
 
-
-class ComponentController:
-    def __init__(self, name, state, mapping, path=()):
+class Controller:
+    def __init__(self, name, remote, mapping={}, path=()):
         self._name = name
-        self._state = state
+        self._remote = remote
         self._mapping = mapping
-        self._remote = Remote(self._state, path=path)
 
     @classmethod
-    def from_config(cls, name, client_config, state, mapping={}, notebook_control=False, **controller_kwargs):
+    def from_config(cls, name, client_config, remote, mapping={}, **controller_kwargs):
         return cls(
             name=name,
-            state=state,
+            remote=remote,
             mapping=mapping,
         )
-        
+
     @property
     def name(self):
         return self._name
@@ -40,7 +37,7 @@ class ComponentController:
         else:
             attr, value = self.to_jax(attr, value)
             setattr(self._remote, attr, value)
-            
+
     def to_jax(self, attr, value=None):
         if attr in self._mapping:
             mapping = self._mapping[attr]
@@ -50,7 +47,10 @@ class ComponentController:
         return (attr, value) if value is not None else attr
 
     def set_state(self, state):
-        self._state = state
+        pass
+    
+    def set_controller_parameters(self, controller_parameters):
+        pass
 
     def fetch_changes(self):
         changes = self._remote.fetch_changes()
@@ -58,4 +58,3 @@ class ComponentController:
 
     def step(self, time, catch_errors):
         pass
-
